pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-reformatter-backend"
version = "0.1.0"
description = "CV reformatting API backend"
requires-python = "==3.12.13"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.41.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "supabase==2.28.3",
    "httpx==0.28.1",
    "python-docx==1.2.0",
    "pypdf==6.10.2",
    "python-multipart==0.0.22",
    "anthropic>=0.34,<1.0",
    "docxtpl>=0.17,<1.0",
    "jinja2>=3.1,<4.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.24.0",
    "ruff==0.8.5",
]

[tool.setuptools]
py-modules = ["models"]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*", "pipeline*", "templates*"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["api", "pipeline", "templates"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
